Modify users account bootstrapping steps
Instead of starting out by using local state, just configure the
backend to use the programmatic credentials for AWSAdministratorAccess
as obtained for the COOL "terraform" account from the AWS SSO page.

Also go ahead and add (but leave commented out) the appropriate line
in backup.tf for this, and comment that file to explain when to use
each "profile =" line.