Add the provisionaccount role name as an output
This seems unnecessary, since the role name is already an input
variable, but we want to specify it here as an output for simple (if
lengthy to explain) reasons.

Suppose we want to attach another policy to this role, and we have the
ARN of this new policy.  The aws_iam_role_policy_attachment Terraform
resource requires a policy ARN and a role name.  If we simply use the
input variable when defining the aws_iam_role_policy_attachment
resource, then the role may not actually exist when Terraform attempts
to create the aws_iam_role_policy_attachment resource.  If we use this
output instead then the role is guaranteed to exist when Terraform
attempts to create the aws_iam_role_policy_attachment resource.