Create a "gods" group that is allowed to assume any role anywhere
This is the start of the work to get rid of the fine-grained groups we
have been creating and move to coarser groups that are defined by job
functions.  This is necessary because AWS has a hard limit of ten for
the number of groups to which a user can belong.